+++ app/repository/RoleRepository.java
@@ -3,8 +3,6 @@ package repository;
 import models.MyRole;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -13,21 +11,23 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class RoleRepository {
-    private static RoleRepository instance;
-    private final Connection connection;
-    private final Logger logger;
 
-    final String TABLE_NAME = "MY_ROLE";
-    final String createTable = "CREATE TABLE IF NOT EXISTS "+ TABLE_NAME +" ("+
-            "ROLE_TYPE varchar(200), "+
-            "ROLE_DESCRIPTION varchar(200) NOT NULL, "+
-            "PRIMARY KEY (ROLE_TYPE))";
+    private final Logger logger = LoggerFactory.getLogger(RoleRepository.class);
+    private static RoleRepository instance = null;
+    private final String TABLE_NAME = "MY_ROLE";
 
     private RoleRepository() {
-        connection = MyDatabase.getConnection();
-        logger = LoggerFactory.getLogger(RoleRepository.class);
+        createTable();
+    }
+
+    private void createTable(){
+        String createTableQuery = "CREATE TABLE IF NOT EXISTS "+ TABLE_NAME +" ("+
+                "ROLE_TYPE varchar(200), "+
+                "ROLE_DESCRIPTION varchar(200) NOT NULL, "+
+                "PRIMARY KEY (ROLE_TYPE))";
         try {
-            PreparedStatement statement = connection.prepareStatement(createTable);
+            Connection connection = MyDatabase.getConnection();
+            PreparedStatement statement = connection.prepareStatement(createTableQuery);
             statement.execute();
             logger.info("Table fetched successfully.");
         } catch (SQLException e) {
@@ -35,43 +35,54 @@ public class RoleRepository {
         }
     }
 
-    public boolean save(@NotNull MyRole role) {
+    public boolean save(MyRole role) {
+        if (!isValidRole(role)) return false;
+        String saveQuery = "INSERT INTO "+TABLE_NAME+ " VALUES(?,?)";
         try {
-            String saveQuery = "INSERT INTO "+TABLE_NAME+ " VALUES(?,?)";
+            Connection connection = MyDatabase.getConnection();
             PreparedStatement statement = connection.prepareStatement(saveQuery);
             statement.setString(1, role.getRoleType());
             statement.setString(2, role.getDescription());
             int count = statement.executeUpdate();
             logger.info(count+" Role saved successfully");
+            statement.close();
+            connection.close();
             return true;
-        } catch (SQLException e) {
+        } catch (SQLException | NullPointerException e) {
             logger.warn(e.getMessage());
         }
         return false;
     }
 
-    public MyRole findUserRoleByType(@NotEmpty @NotNull String roleType) {
+    public MyRole findUserRoleByType(String roleType) {
+        if (roleType == null || roleType.isBlank()) return null;
+        MyRole role = null;
+        String findQuery = "SELECT * FROM "+ TABLE_NAME+" WHERE ROLE_TYPE=?";
         try {
-            String findQuery = "SELECT * FROM "+ TABLE_NAME+" WHERE ROLE_TYPE=?";
+            Connection connection = MyDatabase.getConnection();
             PreparedStatement statement = connection.prepareStatement(findQuery);
             statement.setString(1, roleType);
             ResultSet resultSet = statement.executeQuery();
             if(resultSet.next()) {
-                return MyRole.newBuilder()
+                role = MyRole.newBuilder()
                         .setRoleType(resultSet.getString("ROLE_TYPE"))
                         .setDescription(resultSet.getString("ROLE_DESCRIPTION"))
                         .build();
             }
-        } catch (SQLException e) {
+            resultSet.close();
+            statement.close();
+            connection.close();
+        } catch (SQLException | NullPointerException e) {
             logger.warn(e.getMessage());
         }
-        return null;
+        return role;
     }
 
-    public List<MyRole> findAllUserRoles() {
-        List<MyRole> userRoles = new ArrayList<>();
+    public List<MyRole> findAllRoles() {
+        List<MyRole> roles = new ArrayList<>();
+        String findQuery = "SELECT * FROM "+ TABLE_NAME;
         try {
-            String findQuery = "SELECT * FROM "+ TABLE_NAME;
+            Connection connection = MyDatabase.getConnection();
             PreparedStatement statement = connection.prepareStatement(findQuery);
             ResultSet resultSet = statement.executeQuery();
             while(resultSet.next()) {
@@ -79,40 +90,57 @@ public class RoleRepository {
                         .setRoleType(resultSet.getString("ROLE_TYPE"))
                         .setDescription(resultSet.getString("ROLE_DESCRIPTION"))
                         .build();
-                userRoles.add(role);
+                roles.add(role);
             }
-        } catch (SQLException e) {
+            resultSet.close();
+            statement.close();
+            connection.close();
+        } catch (SQLException | NullPointerException e) {
             logger.warn(e.getMessage());
         }
-        return userRoles;
+        return roles;
     }
 
     public boolean updateUserRoles(MyRole oldRole, MyRole newRole) {
+        if (!(isValidRole(newRole) && isValidRole(oldRole))) return false;
+        String query = "UPDATE "+ TABLE_NAME+ "SET ROLE_TYPE=?, ROLE_DESCRIPTION=? WHERE ROLE_TYPE=?";
         try {
-            String query = "UPDATE "+ TABLE_NAME+ "SET ROLE_TYPE=?, ROLE_DESCRIPTION=? WHERE ROLE_TYPE=?";
+            Connection connection = MyDatabase.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setString(1, newRole.getRoleType());
             statement.setString(2, newRole.getDescription());
             statement.setString(3, oldRole.getRoleType());
             int count = statement.executeUpdate();
+            statement.close();
+            connection.close();
             return (count == 1);
-        } catch (SQLException e) {
+        } catch (SQLException | NullPointerException e) {
             logger.warn(e.getMessage());
         }
         return false;
     }
 
     public boolean delete(MyRole role) {
+        if (!isValidRole(role)) return false;
         try {
             String query = "DELETE FROM "+ TABLE_NAME+ " WHERE ROLE_TYPE =?";
+            Connection connection = MyDatabase.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setString(1, role.getRoleType());
             int count = statement.executeUpdate();
+            statement.close();
+            connection.close();
             return (count == 1);
-        } catch (SQLException e) {
+        } catch (SQLException | NullPointerException e) {
             logger.warn(e.getMessage());
         }
         return false;
+    }
+
+    private boolean isValidRole(MyRole role){
+        return role != null &&
+                !role.getRoleType().isBlank() &&
+                !role.getDescription().isBlank();
     }
 
     public static RoleRepository getInstance(){
